--- src/pages/index.tsx
@@ -1,9 +1,0 @@
-const HomePage = () => {
-  return (
-    <div>
-      <h1>Home Page</h1>
-    </div>
-  )
-}
-
-export default HomePage
